fix(#23): Moves Conditional modal into its own file
* Simplifies Debug.tsx by moving the conditional breakpoint modal component
  into its own file
